updated push_swap

+++ cursus/push_swap/huge_sort.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   huge_sort.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: tcakmako <tcakmako@42kocaeli.com.tr>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/08/03 12:41:41 by tcakmako          #+#    #+#             */
+/*   Updated: 2022/08/03 12:49:14 by tcakmako         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "push_swap.h"
 
 static void	mutual_sort(t_ints *a, t_ints *b)
@@ -36,14 +48,13 @@ void	huge_sort(t_ints *a)
 	size_t	steps_to_sort;
 	size_t	sorting_size;
 
-
 	b = dup_ints(a);
 	steps_to_sort = a->inv / STEP_CONSTANT + 1;
 	sorting_size = a->inv / steps_to_sort;
 	limits[1] = max_element(a);
 	while (steps_to_sort > 0)
 	{
-		if (steps_to_sort == 1 || sorting_size >= limits[1])
+		if (steps_to_sort == 1 || (long) sorting_size >= (long) limits[1])
 			limits[0] = min_element(a);
 		else
 			limits[0] = limits[1] - sorting_size + 1;
